requirements.txt
asgiref==3.8.1
astroid==2.15.8
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.3
django-bootstrap5==24.3
djangorestframework==3.15.2
flake8==6.1.0
flake8-django==1.4
flake8-html==0.4.3
idna==3.10
Jinja2==3.1.4
junit-xml==1.8
lazy-object-proxy==1.10.0
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mutmut==3.2.1
parso==0.8.4
pip-autoremove==0.10.0
platformdirs==4.3.6
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.18.0
python-decouple==3.8
pytz==2024.2
requests==2.32.3
rich==13.9.4
setproctitle==1.3.3
setuptools==75.4.0
six==1.16.0
sqlparse==0.5.1
supervisor==4.2.5
textual==0.85.2
typing_extensions==4.12.2
uc-micro-py==1.0.3
urllib3==2.2.3
wheel==0.45.0
wrapt==1.16.0
